+++ @xen-orchestra/web/src/components/pool/PoolHeader.vue
@@ -2,13 +2,17 @@
   <UiHeadBar :icon="faCity">
     {{ pool.name_label }}
     <template #actions>
-      <UiButton :left-icon="faPlus" variant="secondary" accent="brand" size="medium" @click="goToNewVm()">
+      <UiLink size="medium" :to="{ name: '/vm/new', query: { poolid: pool.id } }" :icon="faPlus">
         {{ t('new-vm') }}
-      </UiButton>
+      </UiLink>
     </template>
   </UiHeadBar>
   <TabList>
-    <TabItem disabled>{{ t('dashboard') }}</TabItem>
+    <RouterLink v-slot="{ isActive, href }" :to="`/pool/${pool.id}/dashboard`" custom>
+      <TabItem :active="isActive" :href tag="a">
+        {{ t('dashboard') }}
+      </TabItem>
+    </RouterLink>
     <TabItem disabled>{{ t('alarms') }}</TabItem>
     <TabItem disabled>{{ t('stats') }}</TabItem>
     <RouterLink v-slot="{ isActive, href }" :to="`/pool/${pool.id}/system`" custom>
@@ -30,7 +34,7 @@
     </RouterLink>
     <RouterLink v-slot="{ isActive, href }" :to="`/pool/${pool.id}/vms`" custom>
       <TabItem :active="isActive" :href tag="a">
-        {{ t('vms') }}
+        {{ t('vms', 2) }}
       </TabItem>
     </RouterLink>
   </TabList>
@@ -40,19 +44,14 @@
 import type { XoPool } from '@/types/xo/pool.type'
 import TabItem from '@core/components/tab/TabItem.vue'
 import TabList from '@core/components/tab/TabList.vue'
-import UiButton from '@core/components/ui/button/UiButton.vue'
 import UiHeadBar from '@core/components/ui/head-bar/UiHeadBar.vue'
+import UiLink from '@core/components/ui/link/UiLink.vue'
 import { faCity, faPlus } from '@fortawesome/free-solid-svg-icons'
 import { useI18n } from 'vue-i18n'
-import { useRouter } from 'vue-router'
 
 const { pool } = defineProps<{
   pool: XoPool
 }>()
 
 const { t } = useI18n()
-
-const router = useRouter()
-
-const goToNewVm = () => router.push({ name: '/vm/new', query: { poolid: pool.id } })
 </script>