+++ @xen-orchestra/web/src/components/vm/dashboard/VmDashboardNetworkUsageChart.vue
@@ -47,7 +47,7 @@ const networkUsage = computed<LinearChartData>(() => {
 
   const rxSeries = [
     {
-      label: t('network-upload'),
+      label: t('network-download'),
       data: timestamps.map((timestamp, index) => ({
         timestamp,
         value: Object.values(data.stats.vifs?.rx ?? {}).reduce((sum, values) => sum + (values[index] ?? NaN), 0),
@@ -57,7 +57,7 @@ const networkUsage = computed<LinearChartData>(() => {
 
   const txSeries = [
     {
-      label: t('network-download'),
+      label: t('network-upload'),
       data: timestamps.map((timestamp, index) => ({
         timestamp,
         value: Object.values(data.stats.vifs?.tx ?? {}).reduce((sum, values) => sum + (values[index] ?? NaN), 0),
@@ -65,7 +65,7 @@ const networkUsage = computed<LinearChartData>(() => {
     },
   ]
 
-  return [...rxSeries, ...txSeries]
+  return [...txSeries, ...rxSeries]
 })
 
 const maxValue = computed(() => {